% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{aggregate_IC}
\alias{aggregate_IC}
\title{Aggregate influence curves to the cluster level}
\usage{
aggregate_IC(recordIC, id)
}
\arguments{
\item{recordIC}{The individual-level IC estimates}

\item{id}{Vector of cluster membership IDs}
}
\value{
A vector of aggregated IC values with length equal to the number of
unique cluster IDs.
}
\description{
Aggregates individual-level influence curve (IC) estimates to the cluster
level. This allows for appropriate statistical inference that respects the
cluster as the independent unit, whether the target parameter is an
individual-level or cluster-level effect.
}
