% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps_functions.R
\name{do.adaptive.prespec}
\alias{do.adaptive.prespec}
\title{Implement adaptive pre-specification}
\usage{
\method{do}{adaptive.prespec}(
  goal,
  target = "indv",
  break.match = T,
  Ldata,
  V = 5,
  cand.QAdj,
  cand.Qform,
  cand.gAdj,
  cand.gform,
  remove.pscore,
  QAdj = NULL,
  gAdj = NULL,
  scale_value,
  scale_value_min,
  verbose = F,
  sample.effect,
  data.adapt.complexity
)
}
\arguments{
\item{goal}{goal of analysis ('aRR', 'OR', 'RD')}

\item{target}{Target of estimation and inference: cluster-level ("clust") or
pooled-indv effect ("indv")}

\item{break.match}{Logical indicating whether to break matched pairs (if they
exist)}

\item{Ldata}{Input data set}

\item{cand.QAdj}{Character vector of names of candidate adjustment variables
for the outcome regression}

\item{cand.Qform}{Character vector of names of candidate adjustment
algorithms for the outcome regression}

\item{cand.gAdj}{Character vector of names of candidate adjustment variables
for the propensity score}

\item{cand.gform}{Character vector of names of candidate adjustment
algorithms for the propensity score}

\item{remove.pscore}{if T, remove the variable(s) selected for adjustment in
the outcome regression from candidates for the pscore... should only be
used if doing adaptive prespec in RCT with few indpt units}

\item{QAdj}{selected adjustment variable for the outcome regression, if
specifying in advannce}

\item{gAdj}{selected adjustment variable for the pscore regression, if
specifying in advance}

\item{scale_value}{maximum value for outcome scaling}

\item{scale_value_min}{minimum value for outcome scaling}

\item{sample.effect}{If the target of inference the sample effect or the
population effect?}

\item{data.adapt.complexity}{Set to 'low', 'med', or 'high'. See
?get.cand.adj for details.}
}
\value{
Working model selection for candidate TMLE
}
\description{
function to implement adaptive prespecification as described in
Balzer et al. "Adaptive pre-specification in randomized trials with and
without pair-matching"
}
