% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_functions.R
\name{do.Init.Qbar}
\alias{do.Init.Qbar}
\title{Function to do initial estimation of E[Y|A,W] = Qbar(A,W)}
\usage{
\method{do}{Init.Qbar}(train, QAdj, Qform = "glm", glm.out = NULL, verbose)
}
\arguments{
\item{train}{Input data.}

\item{QAdj}{Adjustment variable(s)}

\item{Qform}{outcome regression fit}

\item{glm.out}{}

\item{verbose}{Prints more information about what's happening under the
hood.}
}
\value{
List with adjustment variable(s),	outcome regression fit data set
augmented w/ initial predictions: Qbar(A,W), Qbar(1,W) and Qbar(0,W)
}
\description{
Function to do initial estimation of E[Y|A,W] = Qbar(A,W)
}
