% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_functions.R
\name{do.targeting}
\alias{do.targeting}
\title{Function to update initial estimators of QbarAW}
\usage{
\method{do}{targeting}(train, eps, goal)
}
\arguments{
\item{train}{data set}

\item{eps}{fluctuation coefficient}

\item{goal}{'RD' = risk difference; otherwise RR/OR}
}
\value{
data.frame w/ targeted predictions: Qbar*(A,W), Qbar*(1,W), Qbar*(0,W)
}
\description{
Function to update initial estimators of QbarAW
}
