% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_functions.R
\name{do.TMLE}
\alias{do.TMLE}
\title{Obtain point estimates with TMLE}
\usage{
\method{do}{TMLE}(
  goal,
  target,
  train,
  QAdj,
  Qform = "glm",
  gAdj = NULL,
  gform = "glm",
  Q.out = NULL,
  p.out = NULL,
  scale_value,
  scale_value_min,
  doing.CV = F,
  verbose,
  sample.effect
)
}
\arguments{
\item{goal}{'aRR' for arithmetic risk ratio, 'OR' for odds ratio, otherwise
risk difference}

\item{target}{'clust' (cluster-level effect) or 'indv' (pooled
individual-level effect) target of inference; see Stage2 description}

\item{train}{Training data}

\item{QAdj}{Character string of variable names for the adjustment set for the
outcome regression.}

\item{Qform}{Prediction function/algorithm for outcome regression, see Stage2
and/or do.Init.Qbar for options}

\item{gAdj}{Character string of variable names for the adjustment set for the
propensity score.}

\item{gform}{Prediction function/algorithm for propensity score, see Stage2
and/or do.Init.Qbar for options}

\item{Q.out}{Initial estimator of the conditional mean outcome}

\item{p.out}{initial estimator of the propensity score}

\item{scale_value}{For bounded continuous outcomes, the max value in the
input data}

\item{scale_value_min}{For bounded continuous outcomes, the min value in the
input data}

\item{doing.CV}{}

\item{verbose}{Prints more information about what's happening under the
hood.}

\item{sample.effect}{Logical indicating whether to perform inference for the
sample average treatment effect (if TRUE) or the population average
treatment effect (if FALSE).}
}
\value{
A list with: training data augmented with estimates, prespecified
adjustment variables for the conditional mean outcome (QAdj), prespecified
adjustment variables for the propensity score (gAdj), initial estimator of
the conditional mean outcome (Q.out), estimator of the propensity score
(p.out), estimated fluctuation coef (epsilon)
}
\description{
Obtain point estimates with TMLE
}
