% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.cand.adj}
\alias{get.cand.adj}
\title{Generate set of working models for adaptive pre-specification}
\usage{
get.cand.adj(cand.vars, cand.algos = NULL, data.adapt.complexity)
}
\arguments{
\item{cand.vars}{Character vector of variable names that may be used in the
adjustment set. Default is NULL, in which case it returns a single
unadjusted working model.}

\item{cand.algos}{Character vector of algorithm possibilities. See 'Qform' in
Stage2 documentation for supported options. Default is NULL, in which case
only GLMs are returned.}

\item{data.adapt.complexity}{Either 'low' (default), 'med', or 'high'. 'low'
returns GLMs with a single adjustment variable, while 'med' adds all
cand.algos with the full cand.vars specification. 'high' includes all
non-redundant subsets of algorithms and variables, and may lead to
overfitting; use carefully.}
}
\value{
A data frame where each row has a candidate model: The first column
is the algorithm and the second column is a list of variables to use with
it.
}
\description{
Function to get candidate adjustment strategies (variables +
algorithms) for estimating the outcome regression and the propensity score.
}
