% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_functions.R
\name{get.clever.cov}
\alias{get.clever.cov}
\title{Function to calculate the clever covariate}
\usage{
get.clever.cov(train, gAdj, gform, p.out = NULL, verbose)
}
\arguments{
\item{train}{data set}

\item{gAdj}{adjustment variable(s)}

\item{gform}{pscore regression form}

\item{p.out}{}

\item{verbose}{Prints more info}
}
\value{
List with adjustment variable(s), pscore regression, data set
augmented with pscore & clever covariate (H.AW, H.1W, H.0W)
}
\description{
Function to calculate the clever covariate
}
