% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_functions.R
\name{get.epsilon}
\alias{get.epsilon}
\title{Function to calculate the fluctuation coefficient}
\usage{
get.epsilon(train, goal, verbose = F)
}
\arguments{
\item{train}{data set}

\item{goal}{'aRR' = arithmetic RR, 'OR' = odds ratio, 'RD' risk diff}

\item{verbose}{}
}
\value{
Estimated fluctuation coefficient (eps)
}
\description{
Function to calculate the fluctuation coefficient
}
