% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps_functions.R
\name{get.IC.CV}
\alias{get.IC.CV}
\title{Calculate a cross-validated estimate of the influence curve}
\usage{
get.IC.CV(
  goal,
  target,
  break.match,
  verbose,
  Ldata,
  QAdj,
  Qform,
  gAdj = NULL,
  gform = "glm",
  scale_value,
  scale_value_min,
  folds,
  sample.effect
)
}
\arguments{
\item{goal}{goal of analysis ('aRR' or 'RD)}

\item{target}{cluster/individual effect (JN: Not implemented yet?)}

\item{break.match}{indicator to break the match}

\item{Ldata}{input dataset}

\item{QAdj}{adjustment variable for the outcome regression}

\item{Qform}{adjustment approach for outcome regression}

\item{gAdj}{adjustment variable for the pscore}

\item{gform}{adjustment approach for pscore regression}

\item{scale_value}{For outcomes not bounded within \link{0,1}, the maximum value of the outcome.}

\item{scale_value_min}{For outcomes not bounded within \link{0,1}, the minimum value of the outcome.}

\item{folds}{Number of folds in cross-validation.}
}
\value{
cross-validated estimate of the IC for pair
}
\description{
UPDATES
previous version only did leave-one-out (unit or pair)
this version generalizes to V-fold CV if V>=40
\itemize{
\item can input the number of folds V (default=10)
\item folds created stratified on binary outcomes (by default)
\item if stratify=T and # observations in a given class is <V,
then sets V=min observations in that fold
}
}
\details{
See Balzer et al. "Adaptive pre-specification in randomized trials with and without pair-matching."
}
