% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.IC.variance}
\alias{get.IC.variance}
\title{Calculate the variance of the influence curve}
\usage{
get.IC.variance(
  goal,
  target,
  Vdata,
  R1 = NA,
  R0 = NA,
  scale_value = 1,
  scale_value_min = 0,
  doing.CV = F,
  verbose = F,
  sample.effect
)
}
\arguments{
\item{goal}{One of 'aRR'  (arithmetic risk ratio), 'RD' (risk difference), or
'OR' (odds ratio)}

\item{target}{Target of inference, either cluster-level ('clust') or pooled
individivual-level effect ('indv')}

\item{Vdata}{data set}

\item{scale_value}{maximum value for outcome scaling}

\item{scale_value_min}{minimum value for outcome scaling}

\item{doing.CV}{}

\item{sample.effect}{Logical indicating whether to perform inference for the
sample average treatment effect (if TRUE) or the population average
treatment effect (if FALSE).}
}
\value{
on log scale for if goal='aRR' or 'OR' ... estimated IC & variance -
preserving/breaking the match
}
\description{
NOTE: UPDATE OF UNSCALING HAPPENS HERE
}
