% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.inference}
\alias{get.inference}
\title{Calculate confidence intervals and p-values on relative or absolute scale}
\usage{
get.inference(
  goal = "RD",
  psi = NA,
  psi.hat,
  se,
  df = 99,
  sig.level = 0.05,
  one.sided = F,
  alt.smaller = NULL
)
}
\arguments{
\item{goal}{String specifying the scale of the target parameter. Default is
\code{RD}, risk/rate difference. Any other values assume that input values
are given on the log scale, and the function will exponentiate the
estimated target parameter and confidence interval bounds to output an
artihmetic risk/rate ratio.}

\item{psi}{True value (if known) of the target parameter, for example, in a
simulation study.}

\item{psi.hat}{Estimated value of the target parameter.}

\item{se}{Standard error of the estimated target parameter.}

\item{df}{Degrees of freedom for the Student's \emph{t} distribution as an
approximation of the asymptotic normal distribution. If \code{df > 40}, the
value is ignored and a normal distribution is used for inference.}

\item{sig.level}{Desired significance (alpha) level. Defaults to 0.05.}

\item{one.sided}{Logical indicating that a one-sided test is desired.
Defaults to \code{FALSE}.}

\item{alt.smaller}{If one-sided test is desired, is the alternative
hypothesis that the intervention arm will have a smaller value that the
control arm? For example, if you expect a public health intervention to
reduce the mean of a disease outcome, use alt.smaller = TRUE (this
corresponds to a null hypothesis that the intervention did not reduce the
mean disease outcome).}
}
\value{
A one-row data frome with the estimated target parameter value
(\code{est}), the (two-sided) confidence interval \code{CI.lo},
\code{CI/hi}, the standard error of the estimate, the (possibly one-sided)
p-value, and bias/coverage/rejection indicators (if true value or target
parameter is supplied). NOTE: If \code{goal != "RD"}, the output standard
error will be on the log scale.
}
\description{
Calculate confidence intervals and p-values on relative or absolute scale
}
