% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mean.harmonic}
\alias{mean.harmonic}
\title{Harmonic mean}
\usage{
\method{mean}{harmonic}(a)
}
\arguments{
\item{a}{Vector of values}
}
\value{
The harmonic mean of the values: 1 / mean(1/a).
}
\description{
Calculates the harmonic mean of a vector of numbers.
}
