% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{preprocess_and_check_data}
\alias{preprocess_and_check_data}
\title{Check data inputs for Stage2()}
\usage{
preprocess_and_check_data(data.input, verbose)
}
\arguments{
\item{data.input}{Data frame input.}

\item{verbose}{Logical indicating if you want a lot of details to be printed
to the console.}
}
\value{
A data freame with the necessary structure for Stage2().
}
\description{
Performs several checks and throws warnings before running any
statistical analysis.
}
