% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_functions.R
\name{Stage2}
\alias{Stage2}
\title{Two-stage or single-stage RCT estimation}
\usage{
Stage2(
  goal = "aRR",
  target = "indv",
  data.input,
  QAdj = NULL,
  Qform = "glm",
  gAdj = NULL,
  gform = "glm",
  do.data.adapt = F,
  data.adapt.complexity = NULL,
  cand.QAdj = NULL,
  cand.Qform = "glm",
  cand.gAdj = NULL,
  cand.gform = "glm",
  V = 5,
  remove.pscore = NULL,
  do.cv.variance = F,
  sample.effect = T,
  break.match = T,
  one.sided = F,
  alt.smaller = NULL,
  verbose = F,
  psi = NA,
  return.IC = F
)
}
\arguments{
\item{goal}{aRR = arithmetic risk ratio;  RD=risk difference; OR= odds ratio
(not recommended)}

\item{target}{target of inference: cluster-level ("clust") or pooled-indv
effect ("indv"). With appropriate weights, this can recover cluster-level
effects from individual data, or individual-level effects from clustered
data (see note above re: alpha weights).}

\item{data.input}{Observed data with treatment column A, outcome Y. See
further specifications above.}

\item{QAdj}{Character string of variable names to force in to the adjustment
set for the outcome regression. Default is NULL. If non-null, will override
'cand.QAdj'. For an unadjusted outcome regression, specify QAdj = 'U'.}

\item{Qform}{String pre-specifying which algorithm to use to estimate the
outcome regression. Currently implemented options: 'glm', 'lasso', 'mars'
(multivariate adaptive regression splines), 'mars.corP' (MARS that
pre-screens in variables correlated with the outcome), 'step' (stepwise
selection with main terms), and 'step.interaction' (stepwise selection with
interaction terms).}

\item{gAdj}{Character string of variable names to force in to the adjustment
set for the propensity score. Default is NULL. If non-null, will override
'cand.gAdj'. For an unadjusted outcome regression, specify gAdj = 'U'.}

\item{gform}{String pre-specifying which algorithm to use to estimate the
propensity score. Same options as 'Qform'.}

\item{do.data.adapt}{Should adaptive pre-specification (Balzer 2016) be used?
Defaults to FALSE.}

\item{cand.QAdj}{Character vector of candidate adjustment variable(s) for
conditional mean outcome}

\item{cand.Qform}{Character vector with candidate adjustment approache(s) for
conditional mean outcome. See 'Qform' for options.}

\item{cand.gAdj}{Character vector of candidate adjustment variable(s) for
propensity score}

\item{cand.gform}{Character vector with candidate adjustment approache(s) for
propensity score. See 'Qform' for options.}

\item{V}{Number of folds for cross-validation steps. Default is 5, once the
number of independent units exceeds 40. For fewer than 40 independent
units, defaults to LOOCV.}

\item{remove.pscore}{Relevant when do.data.adapt = T. Should adjustment
variables that are selected for the outcome regression be removed from the
candidate set for the propensity score? See 'Collaborative TMLE'. Recommend
remove.pscore = TRUE when there are few independent units; defaults to TRUE
if the number of independent units is under 40.}

\item{do.cv.variance}{Should cross-validated variance estimate be used for
inference? Default FALSE.}

\item{sample.effect}{Is the target the sample effect among the observed
units? If FALSE, target is population effect. Default is TRUE.}

\item{break.match}{If pair-matched trial, should pairs be ignored during
inference?}

\item{one.sided}{Indicator specifying whether a one-sided p-value should be
returned. Defaults to FALSE.}

\item{alt.smaller}{Only needed if one.sided = T. Specifies whether the
alternative hypothesis is that the treatment arm has a lower level of the
outcome. If the outcome is 'bad' (say, incidence of an infectious disease),
and the treatment is hypothesized to be 'good' and reduce incidence,
alt.smaller = T.}

\item{verbose}{Prints more information about what's happening under the
hood.}

\item{psi}{In (say) a simulation study, if the true value of the treatment
effect is known, it can be input here and the outcome will include
an indicator for rejection of the null hypothesis and coverage. Default is
NA.}

\item{return.IC}{indicator of whether to return the influence curve}
}
\value{
Point estimate and inference
}
\description{
Takes data frame input.data and calculates arm-specific and treatment effect estimates.
}
\details{
Outcomes can be bounded continuous (in which case they are scaled to \link{0,1}
before estimation and inference, then unscaled - See Ch7 TLB) or binary. Does
not work for multicategory outcomes.

The input.data object must include treatment indicator column 'A' and outcome
'Y'. If the trial pair-matched and want to keep pairs (break.match = F), the
column with pair IDs must be labeled as 'pair'.

For cluster randomized trials, should also include weights 'alpha', and cluster id 'id'.

column for weights (alpha)
Set = 1 for individually randomized trials.
For cluster randomized trials:
value of the weights depends on the target of inference and data level
Details in Benitez et al. https://arxiv.org/abs/2110.09633v2
let J=number of clusters, N_j = cluster-specific sample size, N_tot = total # participants= sum_j N_j
if target='clust' with cluster-level data, alpha=1
if target='clust' with indv-level data, alpha = 1/N_j
if target='indv' with cluster-level data, alpha = J/N_tot*N_j
if target='indv' with indv-level data, then alpha = 1
for demonstration, see sim2.R in https://github.com/LauraBalzer/Comparing_CRT_Methods
weights must sum to the total # of randomized units

For testing, or for a prespecified (not-adaptive) estimation approach,
specify conditional mean outcome adjustment variables (QAdj) and method
(Qform), and propensity score adjustment variables (gAdj) and method (gform).
Note that this is NOT recommended in general--- instead use Adaptive
Prespecification (do.data.adapt = T).
}
