% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle4rcts-package.R
\docType{package}
\name{tmle4rcts-package}
\alias{tmle4rcts}
\alias{tmle4rcts-package}
\title{tmle4rcts: TMLE for randomized controlled trials}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
