% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_functions.R
\name{get.IC.variance}
\alias{get.IC.variance}
\title{Title}
\usage{
get.IC.variance(
  goal,
  target,
  Vdata,
  R1 = NA,
  R0 = NA,
  sample.effect = T,
  scale_value = 1,
  scale_value_min = 0,
  doing.CV = F
)
}
\arguments{
\item{goal}{}

\item{target}{}

\item{Vdata}{}

\item{R1}{}

\item{R0}{}

\item{sample.effect}{}

\item{scale_value}{}

\item{scale_value_min}{}

\item{doing.CV}{}
}
\description{
Title
}
