% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_functions.R
\name{get.IC.variance}
\alias{get.IC.variance}
\title{Calculate the variance of the influence curve}
\usage{
get.IC.variance(
  goal,
  target,
  Vdata,
  R1 = NA,
  R0 = NA,
  sample.effect = T,
  scale_value = 1,
  scale_value_min = 0,
  doing.CV = F
)
}
\arguments{
\item{goal}{aRR= arithmetic risk ratio; RD for the risk difference; OR for
the odds ratio}

\item{target}{target of inference: cluster-level ("clust") or pooled-indv
effect ("indv") (target)}

\item{Vdata}{data set}

\item{scale_value}{maximum value for outcome scaling}

\item{scale_value_min}{minimum value for outcome scaling}

\item{doing.CV}{}
}
\value{
on log scale for if goal='aRR' or 'OR' ... estimated IC & variance -
preserving/breaking the match
}
\description{
UPDATE OF UNSCALING HAPPENS HERE
}
