% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2_functions.R
\name{get.inference}
\alias{get.inference}
\title{Inference on relative or absolute scale}
\usage{
get.inference(
  goal = "RD",
  psi = NA,
  psi.hat,
  se,
  df = 99,
  sig.level = 0.05,
  one.sided = F,
  alt.smaller = NULL
)
}
\arguments{
\item{goal}{String specifying the scale of the target parameter. Default is
'RD', risk/rate difference. Any other values will give risk/rate ratio.}

\item{psi}{True value (if known) of the target parameter, for example, in a
simulation study.}

\item{psi.hat}{Estimated value of the target parameter.}

\item{se}{Standard error of the estimated target parameter.}

\item{df}{Degrees of freedom (number of independent units - 2).}

\item{sig.level}{Desired significance (alpha) level. Defaults to 0.05.}

\item{one.sided}{Logical indicating that a one-sided test is desired.
Defaults to \code{FALSE}.}

\item{alt.smaller}{If one-sided test is desired, ...}
}
\value{
A one-row data frome with the estimated target parameter value
(\code{est}), the (two-sided) confidence interval \code{CI.lo},
\code{CI/hi}, the standard error of the estimate, the (possibly one-sided)
p-value, and bias/coverage/rejection indicators (if true value or target
parameter is supplied).
}
\description{
Inference on relative or absolute scale
}
