% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_me_out}
\alias{test_me_out}
\title{Test Me Out}
\usage{
test_me_out(x)
}
\arguments{
\item{x}{The input object}
}
\value{
A character string.
}
\description{
A test function made while working out package and github setup.
}
\examples{
test_me_out(1)
test_me_out("Yes")
}
